requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.13
langchain-openai>=0.0.2
faiss-cpu>=1.7.4
openai>=1.8.0
python-dotenv>=1.0.0
tiktoken==0.9.0

# Web interface
streamlit>=1.31.0

# Remove unnecessary dependencies
# python-docx==1.1.2
# unstructured[all-docs]

# Remove these as they're either included in other packages or causing issues
# langchain-unstructured  # included in unstructured[all-docs]
# python-magic-bin  # causing platform compatibility issues 